+++ authz/client.go
@@ -28,9 +28,84 @@ var (
 	ErrMissingCaller           = errors.New("missing caller")
 	ErrMissingSubject          = errors.New("missing subject")
 
-	checkResponseDenied  = claims.CheckResponse{Allowed: false}
-	checkResponseAllowed = claims.CheckResponse{Allowed: true}
+	checkResponseDenied  = CheckResponse{Allowed: false}
+	checkResponseAllowed = CheckResponse{Allowed: true}
 )
+
+// CheckRequest describes the requested access.
+// This is designed bo to play nicely with the kubernetes authorization system:
+// https://github.com/kubernetes/kubernetes/blob/v1.30.3/staging/src/k8s.io/apiserver/pkg/authorization/authorizer/interfaces.go#L28
+type CheckRequest struct {
+	// The requested access verb.
+	// this includes get, list, watch, create, update, patch, delete, deletecollection, and proxy,
+	// or the lowercased HTTP verb associated with non-API requests (this includes get, put, post, patch, and delete)
+	Verb string
+
+	// API group (dashboards.grafana.app)
+	Group string
+
+	// ~Kind eg dashboards
+	Resource string
+
+	// tenant isolation
+	Namespace string
+
+	// The specific resource
+	// In grafana, this was historically called "UID", but in k8s, it is the name
+	Name string
+
+	// Optional subresource
+	Subresource string
+
+	// For non-resource requests, this will be the requested URL path
+	Path string
+
+	// Folder is the parent folder of the requested resource
+	Folder string
+}
+
+type CheckResponse struct {
+	// Allowed is true if the request is allowed, false otherwise.
+	Allowed bool
+}
+
+type AccessChecker interface {
+	// Check checks whether the user can perform the given action for all requests
+	Check(ctx context.Context, id claims.AuthInfo, req CheckRequest) (CheckResponse, error)
+}
+
+type ListRequest struct {
+	// The requested access verb.
+	// this includes get, list, watch, create, update, patch, delete, deletecollection, and proxy,
+	// or the lowercased HTTP verb associated with non-API requests (this includes get, put, post, patch, and delete)
+	Verb string
+
+	// API group (dashboards.grafana.app)
+	Group string
+
+	// ~Kind eg dashboards
+	Resource string
+
+	// tenant isolation
+	Namespace string
+
+	// Optional subresource
+	Subresource string
+}
+
+// TODO: Should the namespace be specified in the request instead.
+// I don't think we'll be able to Compile over multiple namespaces.
+// Checks access while iterating within a resource
+type ItemChecker func(namespace string, name string) bool
+
+type AccessClient interface {
+	AccessChecker
+
+	// Compile generates a function to check whether the id has access to items matching a request
+	// This is particularly useful when you want to verify access to a list of resources.
+	// Returns nil if there is no access to any matching items
+	Compile(ctx context.Context, id claims.AuthInfo, req ListRequest) (ItemChecker, error)
+}
 
 type ClientConfig struct {
 	// RemoteAddress is the address of the authz service. It should be in the format "host:port".
@@ -43,7 +118,7 @@ type ClientConfig struct {
 
 // ClientImpl will implement the claims.AccessClient interface
 // Once we are able to deal with folder permissions expansion.
-var _ claims.AccessChecker = (*ClientImpl)(nil)
+var _ AccessChecker = (*ClientImpl)(nil)
 
 type LegacyClientOption func(*ClientImpl)
 
@@ -159,7 +234,7 @@ func NewClient(cfg *ClientConfig, opts ...LegacyClientOption) (*ClientImpl, erro
 // Implementation
 // -----
 
-func validateAccessRequest(req claims.CheckRequest) error {
+func validateAccessRequest(req CheckRequest) error {
 	if req.Namespace == "" {
 		return ErrMissingRequestNamespace
 	}
@@ -181,7 +256,7 @@ func validateAccessRequest(req claims.CheckRequest) error {
 	return nil
 }
 
-func (c *ClientImpl) check(ctx context.Context, id claims.AuthInfo, req *claims.CheckRequest) (bool, error) {
+func (c *ClientImpl) check(ctx context.Context, id claims.AuthInfo, req *CheckRequest) (bool, error) {
 	ctx, span := c.tracer.Start(ctx, "ClientImpl.hasAccess")
 	defer span.End()
 
@@ -219,7 +294,7 @@ func (c *ClientImpl) check(ctx context.Context, id claims.AuthInfo, req *claims.
 }
 
 // HasAccess implements claims.AccessClient.
-func (c *ClientImpl) Check(ctx context.Context, id claims.AuthInfo, req claims.CheckRequest) (claims.CheckResponse, error) {
+func (c *ClientImpl) Check(ctx context.Context, id claims.AuthInfo, req CheckRequest) (CheckResponse, error) {
 	ctx, span := c.tracer.Start(ctx, "ClientImpl.Check")
 	defer span.End()
 
@@ -302,7 +377,7 @@ func (c *ClientImpl) Check(ctx context.Context, id claims.AuthInfo, req claims.C
 	}
 
 	// Check if the user has access to any of the requested resources
-	return claims.CheckResponse{Allowed: res}, nil
+	return CheckResponse{Allowed: res}, nil
 }
 
 func (c *ClientImpl) validateCaller(caller claims.AuthInfo) error {
@@ -355,7 +430,7 @@ func newOutgoingContext(ctx context.Context) context.Context {
 // CACHE
 // -----
 
-func checkCacheKey(subj string, req *claims.CheckRequest) string {
+func checkCacheKey(subj string, req *CheckRequest) string {
 	return fmt.Sprintf("check-%s-%s-%s-%s-%s-%s-%s-%s-%s", req.Namespace, subj, req.Group, req.Resource, req.Verb, req.Name, req.Subresource, req.Path, req.Folder)
 }
